+++ tests/python/test_workitems.py
@@ -3,9 +3,12 @@ import pytest
 from RPA.Robocloud.Items import BaseAdapter, Items
 
 
+VARIABLES_FIRST = {"username": "testguy", "address": "guy@company.com"}
+VARIABLES_SECOND = {"username": "another", "address": "dude@company.com"}
+
 VALID_DATABASE = {
-    ("test-ws", "test-item"): {"username": "testguy", "address": "guy@company.com"},
-    ("test-ws", "second-item"): {"username": "another", "address": "dude@company.com"},
+    ("workspace-id", "workitem-id-first"): {"variables": VARIABLES_FIRST},
+    ("workspace-id", "workitem-id-second"): {"variables": VARIABLES_SECOND},
 }
 
 
@@ -27,11 +30,13 @@ class MockAdapter(BaseAdapter):
 
 @pytest.fixture
 def valid_adapter(monkeypatch):
-    monkeypatch.setenv("RC_WORKSPACE_ID", "test-ws")
-    monkeypatch.setenv("RC_WORKITEM_ID", "test-item")
+    monkeypatch.setenv("RC_WORKSPACE_ID", "workspace-id")
+    monkeypatch.setenv("RC_WORKITEM_ID", "workitem-id-first")
     MockAdapter.DATABASE = copy.deepcopy(VALID_DATABASE)
-    yield MockAdapter
-    MockAdapter.DATABASE = {}
+    try:
+        yield MockAdapter
+    finally:
+        MockAdapter.DATABASE = {}
 
 
 def test_no_env(monkeypatch):
@@ -51,7 +56,7 @@ def test_load_env(valid_adapter):
     # Work item loaded using env variables
     env = lib.current
     assert env is not None
-    assert env.data["username"] == "testguy"
+    assert env.data["variables"] == VARIABLES_FIRST
 
 
 def test_load_env_disable(valid_adapter):
@@ -62,25 +67,27 @@ def test_load_env_disable(valid_adapter):
     assert lib.current is None
 
 
-def test_keyword_load_item(valid_adapter):
+def test_keyword_load_work_item(valid_adapter):
     lib = Items(default_adapter=valid_adapter)
 
-    item = lib.load_work_item("test-ws", "second-item")
-    assert item.data["username"] == "another"
+    item = lib.load_work_item("workspace-id", "workitem-id-second")
+    assert item.data["variables"] == VARIABLES_SECOND
     assert item == lib.current
 
 
-def test_keyword_save_item(valid_adapter):
+def test_keyword_save_work_item(valid_adapter):
     lib = Items(default_adapter=valid_adapter)
-    item = lib.load_work_item("test-ws", "second-item")
-    MockAdapter.validate(item, "username", "another")
+    item = lib.load_work_item("workspace-id", "workitem-id-second")
+    MockAdapter.validate(item, "variables", VARIABLES_SECOND)
 
-    item.data["username"] = "changed"
+    modified = {"username": "changed", "address": "dude@company.com"}
+    item.data["variables"] = modified
+
     lib.save_work_item()
-    MockAdapter.validate(item, "username", "changed")
+    MockAdapter.validate(item, "variables", modified)
 
 
-def test_keyword_no_active_item():
+def test_no_active_item():
     lib = Items(default_adapter=MockAdapter)
     assert lib.current is None
 
@@ -88,3 +95,72 @@ def test_keyword_no_active_item():
         lib.save_work_item()
 
     assert str(err.value) == "No active work item"
+
+
+def test_list_variables(valid_adapter):
+    lib = Items(default_adapter=valid_adapter)
+    lib.load_work_item("workspace-id", "workitem-id-second")
+
+    names = lib.list_work_item_variables()
+
+    assert len(names) == 2
+    assert "username" in names
+    assert "address" in names
+
+
+def test_delete_variables(valid_adapter):
+    lib = Items(default_adapter=valid_adapter)
+    lib.load_work_item("workspace-id", "workitem-id-second")
+
+    assert "username" in lib.list_work_item_variables()
+    lib.delete_work_item_variables("username")
+    assert "username" not in lib.list_work_item_variables()
+
+    lib.delete_work_item_variables("doesntexist")
+
+    with pytest.raises(KeyError):
+        lib.delete_work_item_variables("doesntexist", force=False)
+
+
+def test_delete_variables_multiple(valid_adapter):
+    lib = Items(default_adapter=valid_adapter)
+    lib.load_work_item("workspace-id", "workitem-id-second")
+
+    assert "username" in lib.list_work_item_variables()
+    assert len(lib.current["variables"]) == 2
+
+    lib.delete_work_item_variables("username")
+
+    assert "username" not in lib.list_work_item_variables()
+    assert len(lib.current["variables"]) == 1
+
+
+def test_delete_variables_multiple(valid_adapter):
+    lib = Items(default_adapter=valid_adapter)
+    lib.load_work_item("workspace-id", "workitem-id-second")
+
+    names = lib.list_work_item_variables()
+    assert "username" in names
+    assert "address" in names
+    assert len(names) == 2
+
+    lib.delete_work_item_variables("username", "address")
+
+    names = lib.list_work_item_variables()
+    assert "username" not in names
+    assert "username" not in names
+    assert len(names) == 0
+
+
+def test_delete_variables_unknown(valid_adapter):
+    lib = Items(default_adapter=valid_adapter)
+    lib.load_work_item("workspace-id", "workitem-id-second")
+
+    assert len(lib.list_work_item_variables()) == 2
+
+    lib.delete_work_item_variables("unknown-variable")
+    assert len(lib.list_work_item_variables()) == 2
+
+    with pytest.raises(KeyError):
+        lib.delete_work_item_variables("unknown-variable", force=False)
+    assert len(lib.list_work_item_variables()) == 2